Add a new script to analyse the number of seconds between each commit

+++ forceResetPeriodically.sh
@@ -1,8 +1,23 @@
 #!/bin/bash
 
+function logToDevAnalysisFile {
+	echo $secondsPassed >> $currentAnalysisFile
+}
+
+function createDevAnalysisFile {
+	mkdir -p $DEV_ANALYSIS_REPO
+	currentAnalysisFile="$DEV_ANALYSIS_REPO/analysisFile-$(eval date \"+%Y%m%d%H%M%S\").log"
+	touch $currentAnalysisFile
+}
+
 function gitResetHard {
-	git reset --hard HEAD
-    git clean -fd
+	if [ "$FAKEIT" = "1" ]
+	then
+		echo "GitResetHard"
+	else
+		git reset --hard HEAD
+		git clean -fd
+	fi
 }
 
 function displayRemainingSecond {
@@ -21,6 +36,7 @@ function displayRemainingSecond {
 
 function resetTimer {
 	timer=$originalTimer
+	secondsPassed=0
 	echo "Timer resetted"
 }
 
@@ -29,40 +45,55 @@ function resetIfLastCommitChanged {
 	if [ "$localLastCommit" != "$lastCommit" ]
 	then
 		echo -ne "\nLast commit changed from $lastCommit to $localLastCommit. "
+		logToDevAnalysisFile 
 		lastCommit=$localLastCommit
 		resetTimer
 	fi
 }
+
 
 function loopInfinitly {
 	while :; do
 		displayRemainingSecond
 		resetIfLastCommitChanged
 		timer=$(($timer - 1))
+		secondsPassed=$(($secondsPassed + 1))
 		sleep 1
 		if [ $timer = 0 ]
 		then
 			resetTimer
-			echo "gitResetHard"
+			gitResetHard
 		fi
 	done
 }
 
+function init {
+	if [ "$numberOfMinutes" = "" ]
+	then
+		echo "No minutes provided. Using 2 minutes as a default timer"
+		numberOfMinutes=2
+	fi
+	echo -e "Timer set to $numberOfMinutes minutes. Have fun !"
+
+	originalTimer=$(($numberOfMinutes * 60))
+	resetTimer
+	secondsPassed=0
+
+	# Global last commit var.
+	# If the commit changed, we reset the timer
+	getLastCommitCommand="git log --format=oneline | head -n 1 | cut -f 1 -d \" \""
+	lastCommit=$(eval $getLastCommitCommand)
+
+	createDevAnalysisFile
+}
+
+DEV_ANALYSIS_REPO=".devAnalysis"
 numberOfMinutes=$1
-
-if [ "$numberOfMinutes" = "" ]
+FAKEIT=$2
+if [ "$FAKEIT" = "1" ]
 then
-	echo "No minutes provided. Using 2 minutes as a default timer"
-	numberOfMinutes=2
+	echo "FAKE IT MODE ACTIVATED !!"
 fi
-echo -e "Timer set to $numberOfMinutes minutes. Have fun !"
 
-originalTimer=$(($numberOfMinutes * 60))
-resetTimer
-
-# Global last commit var.
-# If the commit changed, we reset the timer
-getLastCommitCommand="git log --format=oneline | head -n 1 | cut -f 1 -d \" \""
-lastCommit=$(eval $getLastCommitCommand)
-
+init
 loopInfinitly
